pyproject.toml
[tool.poetry]
name = "phasenet-pytorch"
version = "0.0.1"
description = "PyTorch adaption of PhaseNet"
authors = [
    "Ziyi Xi <ziyixi@mail.ustc.edu.cn>",
    "Weiqiang Zhu <wayne.weiqiang@gmail.com>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
hydra-core = "^1.1.1"
omegaconf = "^2.1.1"
torch = "^1.11.0"
pyasdf = "^0.7.5"
torchvision = "^0.12.0"
obspy = "^1.3.0"
tqdm = "^4.63.1"
matplotlib = "^3.5.1"
torchaudio = "^0.11.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
pylint = "^2.13.1"
autopep8 = "^1.6.0"
Sphinx = "^4.4.0"
rstcheck = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
